Resolve build/type cast errors

+++ rtos/TARGET_CORTEX/TOOLCHAIN_ARM_STD/mbed_boot_arm_std.c
@@ -28,17 +28,17 @@ __value_in_regs struct __argc_argv __rt_lib_init(unsigned heapbase, unsigned hea
 void _platform_post_stackheap_init(void);
 
 #if !defined(ISR_STACK_START)
-extern uint32_t Image$$ARM_LIB_STACK$$ZI$$Base[];
-extern uint32_t Image$$ARM_LIB_STACK$$ZI$$Length[];
+extern uint32_t               Image$$ARM_LIB_STACK$$ZI$$Base[];
+extern uint32_t               Image$$ARM_LIB_STACK$$ZI$$Length[];
 #define ISR_STACK_START       Image$$ARM_LIB_STACK$$ZI$$Base
 #define ISR_STACK_SIZE        Image$$ARM_LIB_STACK$$ZI$$Length
 #endif
 
 #if !defined(HEAP_START)
 // Heap here is considered starting after ZI ends to Stack start
-extern uint32_t Image$$RW_IRAM1$$ZI$$Limit[];
+extern uint32_t               Image$$RW_IRAM1$$ZI$$Limit[];
 #define HEAP_START            Image$$RW_IRAM1$$ZI$$Limit
-#define HEAP_SIZE             ((uint32_t)(ISR_STACK_START - HEAP_START))
+#define HEAP_SIZE             ((uint32_t)((uint32_t)ISR_STACK_START - (uint32_t)HEAP_START))
 #endif
 
 /*
